requirements.txt
google-generativeai
streamlit
Pillow
python-dotenv
requests
plotly
streamlit-lottie

-e .